requirements.txt
absl-py==1.0.0
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
attrs==22.1.0
audioread==2.1.9
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
Click==7.0
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
dill==0.3.3
distlib==0.3.2
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dlib==19.24.0
docutils==0.17.1
experta==1.9.4
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.49.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.0.12
flatbuffers==2.0
fonttools==4.31.2
frozendict==1.2
future==0.18.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.9.0
h5py==3.6.0
httptools==0.1.1
idna==2.10
ikpy==3.2.2
imageio==2.20.0
imagesize==1.2.0
importlib-metadata==4.5.0
ipython==7.34.0
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.2
libclang==13.0.0
librosa==0.9.2
llvmlite==0.38.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.10.1
mpi==1.0.0
mpi-map==2.6
mpi4py==3.0.3
mpi4py-utilities==1.0.1
mpmath==1.2.1
mtcnn==0.1.1
mysqlclient==2.0.3
networkx==2.6.3
numba==0.55.2
numpy==1.21.6
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-contrib-python-headless==4.2.0.32
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pooch==1.6.0
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.2.2
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.4
pydub==0.25.1
pyfcm==1.5.4
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.8
schema==0.6.7
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
starlette==0.12.9
sympy==1.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==3.10.0.0
urllib3==1.26.10
uvicorn==0.11.3
virtualenv==20.4.7
wcwidth==0.2.5
websockets==8.1
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.4.1
